getResourceMCR
