+++ src/pages/memos/[functionId]/index.js
@@ -202,8 +202,7 @@ const Financial = () => {
         access,
         functionId: functionId,
         getEndpoint,
-        getGLResourceId,
-        getResourceMCR
+        getGLResourceId
       },
       width: 1000,
       height: 670,
@@ -227,23 +226,6 @@ const Financial = () => {
     })
     invalidate()
     toast.success(platformLabels.Deleted)
-  }
-
-
-    const getResourceMCR = functionId => {
-    const fn = Number(functionId)
-    switch (fn) {
-      case SystemFunction.CreditNote:
-        return ResourceIds.MCRCreditNote
-      case SystemFunction.DebitNote:
-        return ResourceIds.MCRDebitNote
-      case SystemFunction.ServiceBill:
-        return ResourceIds.MCRServiceBillReceived
-      case SystemFunction.ServiceInvoice:
-        return ResourceIds.MCRServiceInvoice
-      default:
-        return null
-    }
   }
 
   return (
