+++ src/pages/sa-trx/[functionId]/index.js
@@ -223,8 +223,7 @@ const SaTrx = () => {
         functionId,
         lockRecord,
         getResourceId,
-        getGLResource,
-        getResourceMCR
+        getGLResource
       },
       width: 1330,
       height: 720,
@@ -269,22 +268,6 @@ const SaTrx = () => {
     })
     invalidate()
     toast.success(platformLabels.Deleted)
-  }
-
-  const getResourceMCR = functionId => {
-    const fn = Number(functionId)
-    switch (fn) {
-      case SystemFunction.SalesInvoice:
-        return ResourceIds.MCRSalesInvoice
-      case SystemFunction.SalesReturn:
-        return ResourceIds.MCRSalesReturn
-      case SystemFunction.ConsignmentIn:
-        return ResourceIds.MCRClientGOCIn
-      case SystemFunction.ConsignmentOut:
-        return ResourceIds.MCRClientGOCOut
-      default:
-        return null
-    }
   }
 
   return (
